Refactor callback
