+++ src/anomalib/trainer/callbacks/__init__.py
@@ -1,0 +1,8 @@
+"""Necessary callbacks for Trainer."""
+
+# Copyright (C) 2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .post_processor import PostProcessorCallback
+
+__all__ = ["PostProcessorCallback"]
